Make custom menu openable from side of chat
